fix ui of connecting devices screen
